+++ application/models/Tasks.php
@@ -58,6 +58,28 @@ class Tasks extends CSV_Model {
             return $config;
         }
 
+        // Render the current DTO
+        private function showit()
+        {
+            $this->load->helper('form');
+            $task = $this->session->userdata('task');
+            $this->data['id'] = $task->id;
+
+            // if no errors, pass an empty message
+            if ( ! isset($this->data['error']))
+                $this->data['error'] = '';
+
+            $fields = array(
+                'ftask'      => form_label('Task description') . form_input('task', $task->task),
+                'fpriority'  => form_label('Priority') . form_dropdown('priority', $this->app->priority(), $task->priority),
+                'zsubmit'    => form_submit('submit', 'Update the TODO task'),
+            );
+            $this->data = array_merge($this->data, $fields);
+
+            $this->data['pagebody'] = 'itemedit';
+            $this->render();
+        }
+
         // Initiate adding a new task
         public function add()
         {